Refactor password strength Bar Directive
