+++ generators/client-2/templates/src/main/webapp/app/account/_account.module.ts
@@ -4,6 +4,7 @@ import { Password } from './password/password.service';
 import { PasswordResetInit } from './reset/request/reset-request.service';
 import { PasswordResetFinish } from './reset/finish/reset-finish.service';
 
+import { PasswordStrengthBarComponent} from "./password/password-strength-bar.component";
 import { upgradeAdapter } from "../upgrade_adapter";
 
 import { RegisterStateConfig } from "./register/register.state";
@@ -26,7 +27,8 @@ angular
     ])
     .config(AccountStateConfig)
     .config(RegisterStateConfig)
-    .directive('register', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(RegisterComponent))
+    .controller('RegisterController', RegisterController)
+    .directive('passwordStrengthBar', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(PasswordStrengthBarComponent))
     .factory('Register', Register)
     .factory('Activate', Activate)
     .factory('Password', Password)
