+++ generators/client-2/templates/src/main/webapp/app/account/password/password.html
@@ -34,7 +34,7 @@
                             Your password cannot be longer than 50 characters.
                         </p>
                     </div>
-                    <password-strength-bar password-to-check="vm.password"></password-strength-bar>
+                    <password-strength-bar [password-to-check]="vm.password"></password-strength-bar>
                 </div>
                 <div class="form-group">
                     <label class="control-label" for="confirmPassword" translate="global.form.confirmpassword">New password confirmation</label>
