+++ generators/client-2/templates/src/main/webapp/app/account/reset/finish/reset-finish.html
@@ -43,7 +43,7 @@
                                 Your password cannot be longer than 50 characters.
                             </p>
                         </div>
-                        <password-strength-bar password-to-check="vm.resetAccount.password"></password-strength-bar>
+                        <password-strength-bar [password-to-check]="vm.resetAccount.password"></password-strength-bar>
                     </div>
 
                     <div class="form-group">
